+++ lsql-cli/src/main/java/com/w11k/lsql/cli/java/DataClassMeta.java
@@ -59,7 +59,7 @@ public final class DataClassMeta {
             this.fieldType = fieldType;
         }
 
-        public String getColumnsJavaCodeName() {
+        public String getColumnJavaCodeName() {
             return columnsJavaCodeName;
         }