+++ lsql-core/src/main/java/com/w11k/lsql/TypedTable.java
@@ -8,6 +8,17 @@ import java.util.function.Supplier;
 
 public abstract class TypedTable<T extends TableRow, I> {
 
+    private static Table.RowKeyHandler ROW_KEY_HANDLER = new Table.RowKeyHandler() {
+        @Override
+        public String sqlToJava(String internalSql) {
+            return internalSql;
+        }
+
+        @Override
+        public String javaToSql(String javaIdentifier) {
+            return javaIdentifier;
+        }
+    };
     private Supplier<Table> table;
 
     public TypedTable(LSql lSql, String tableName, Class<T> tableRowClass) {
@@ -31,7 +42,7 @@ public abstract class TypedTable<T extends TableRow, I> {
         // Remove null values so that the DB can insert the default values
         map.entrySet().removeIf(entry -> entry.getValue() == null);
 
-        return (Optional<I>) this.table.get().insert(new Row(map));
+        return (Optional<I>) this.table.get().insert(new Row(map), ROW_KEY_HANDLER);
     }
 
     public T insertAndLoad(T instance) {
@@ -44,7 +55,7 @@ public abstract class TypedTable<T extends TableRow, I> {
     }
 
     public Optional<T> load(I id) {
-        Optional<LinkedRow> internalRow = this.table.get().load(id);
+        Optional<LinkedRow> internalRow = this.table.get().load(id, ROW_KEY_HANDLER);
 
         if (!internalRow.isPresent()) {
             return Optional.absent();
@@ -61,21 +72,21 @@ public abstract class TypedTable<T extends TableRow, I> {
 
     public void delete(T instance) {
         Map<String, Object> map = instance.toInternalMap();
-        this.table.get().delete(new Row(map));
+        this.table.get().delete(new Row(map), ROW_KEY_HANDLER);
     }
 
     public void deleteById(I id) {
-        this.table.get().delete(id);
+        this.table.get().delete(id, ROW_KEY_HANDLER);
     }
 
     public void update(T instance) {
         Map<String, Object> map = instance.toInternalMap();
-        this.table.get().update(new Row(map));
+        this.table.get().update(new Row(map), ROW_KEY_HANDLER);
     }
 
     public void updateWhere(T instance, Map<String, Object> where) {
         Map<String, Object> map = instance.toInternalMap();
-        this.table.get().updateWhere(new Row(map), new Row(where));
+        this.table.get().updateWhere(new Row(map), new Row(where), ROW_KEY_HANDLER);
     }
 
     public T updateWith(I id, Function<T, T> with) {
@@ -88,7 +99,7 @@ public abstract class TypedTable<T extends TableRow, I> {
     @SuppressWarnings("unchecked")
     public Optional<I> save(T instance) {
         Map<String, Object> map = instance.toInternalMap();
-        return (Optional<I>) this.table.get().save(new Row(map));
+        return (Optional<I>) this.table.get().save(new Row(map), ROW_KEY_HANDLER);
     }
 
     @SuppressWarnings("unchecked")