+++ lsql-core/src/test/java/com/w11k/lsql/tests/SqlStatementTest.java
@@ -7,7 +7,7 @@ import com.w11k.lsql.converter.Converter;
 import com.w11k.lsql.dialects.PostgresDialect;
 import com.w11k.lsql.dialects.RowKeyConverter;
 import com.w11k.lsql.exceptions.QueryException;
-import com.w11k.lsql.query.RowQuery;
+import com.w11k.lsql.query.PlainQuery;
 import com.w11k.lsql.statement.AbstractSqlStatement;
 import com.w11k.lsql.statement.SqlStatementToPreparedStatement;
 import org.testng.Assert;
@@ -213,7 +213,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     @Test
     public void statementThreeParametersMultipleOccurences() {
         setup();
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement("SELECT * FROM person WHERE " +
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement("SELECT * FROM person WHERE " +
                 "id > /*val=*/ 99999 /**/" +
                 "AND age > /*val=*/ 99999 /**/" +
                 "AND fullname= /*=*/ 'c' /**/" +
@@ -230,7 +230,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
         setup();
         lSql.executeRawSql("delete from person;");
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "insert into person (id, age) \n" +
                         "values ( /*id=*/ 100 /**/,/*age=*/ 100 /**/);"
         );
@@ -245,7 +245,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     @Test
     public void statementQueryParameter() {
         setup();
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement("SELECT * FROM \n" +
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement("SELECT * FROM \n" +
                 "person \n" +
                 "WHERE \n" +
                 "id = /*=*/ 99999 /**/\n" +
@@ -263,7 +263,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void nullValueAsQueryParameter() {
         createTable();
         insert(0, 50, null);
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "SELECT * FROM person WHERE fullname IS /*: string =*/ NULL /**/;");
 
         List<Row> rows = statement.query("fullname", null).toList();
@@ -274,7 +274,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void nullValueAsNamedQueryParameter() {
         createTable();
         insert(0, 50, null);
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "SELECT * FROM person WHERE fullname IS /*aaa: string =*/ NULL /**/;");
 
         List<Row> rows = statement.query("aaa", null).toList();
@@ -286,7 +286,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
         setup();
         List<Row> rows;
 
-        AbstractSqlStatement<RowQuery> statement = this.lSql.createSqlStatement("select * from person where " +
+        AbstractSqlStatement<PlainQuery> statement = this.lSql.createSqlStatement("select * from person where " +
                 "age in (/*ages=*/ 11, 12, 13 /**/) " +
                 "and 1 = /*param=*/ 1 /**/;");
 
@@ -341,7 +341,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
         setup();
         List<Row> rows;
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement("select * from person where" +
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement("select * from person where" +
                 " age in (/*ages=*/ 11, 12, 13 /**/) " +
                 "and 1 = /*param=*/ 1 /**/;");
 
@@ -365,7 +365,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void listLiteralQueryParameterWithPlaceholderStrings() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement("select * from person where" +
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement("select * from person where" +
                 " fullname in (/*fullnames=*/ 'a', 'b', 'c' /**/) ;");
 
         List<Row> rows = statement.query("fullnames", ListLiteralQueryParameter.of("a", "x")).toList();
@@ -382,7 +382,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
         setup();
         List<Row> rows;
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement("select * from person where" +
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement("select * from person where" +
                 " age in (/*ages=*/ 11, 12, 13 /**/) " +
                 "and 1 = /*param=*/ 1 /**/;");
 
@@ -398,9 +398,9 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void resultSetTypeAnnotations() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select count(id) as count_id /*:string*/, max(age) as max_age /*:long*/ from person;");
-        RowQuery query = statement.query();
+        PlainQuery query = statement.query();
         List<Row> rows = query.toList();
         assertEquals(rows.size(), 1);
         Row row = rows.get(0);
@@ -412,9 +412,9 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void resultSetTypeAnnotationsWithQuotedIdentifier() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select count(id) as \"count_id\" /*:string*/, max(age) as \"max_age\" /*:long*/ from person;");
-        RowQuery query = statement.query();
+        PlainQuery query = statement.query();
         List<Row> rows = query.toList();
         assertEquals(rows.size(), 1);
         Row row = rows.get(0);
@@ -426,7 +426,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void explicitParameterConverter() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select * from person where id = /*=*/ 1 /**/;")
                 .setParameterConverter("id", new Converter(String.class, Types.INTEGER) {
                     @Override
@@ -449,7 +449,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void queryParameterTypeAnnotations1() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select * from person where age > /*:int=*/ 18 /**/;");
 
         SqlStatementToPreparedStatement.Parameter param = statement.getParameters().get("age").get(0);
@@ -460,7 +460,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void queryParameterTypeAnnotations2() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select * from person where age > /* :  int  =*/ 18 /**/;");
 
         SqlStatementToPreparedStatement.Parameter param = statement.getParameters().get("age").get(0);
@@ -471,7 +471,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void queryParameterTypeAnnotations3() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select * from person where age > /*p1:int=*/ 18 /**/;");
 
         SqlStatementToPreparedStatement.Parameter param = statement.getParameters().get("p1").get(0);
@@ -482,7 +482,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void queryParameterTypeAnnotations4() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select * from person where age > /* p1 : int =*/ 18 /**/;");
 
         SqlStatementToPreparedStatement.Parameter param = statement.getParameters().get("p1").get(0);
@@ -493,7 +493,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void queryParameterTypeAnnotations5() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select * from person where age > /*p1: int =*/ 18 /**/;");
 
         SqlStatementToPreparedStatement.Parameter param = statement.getParameters().get("p1").get(0);
@@ -504,7 +504,7 @@ public class SqlStatementTest extends AbstractLSqlTest {
     public void queryParameterTypeAnnotationsAreCheckedAtQueryTime() {
         setup();
 
-        AbstractSqlStatement<RowQuery> statement = lSql.createSqlStatement(
+        AbstractSqlStatement<PlainQuery> statement = lSql.createSqlStatement(
                 "select * from person where age > /*:int =*/ 18 /**/;");
 
         statement.query("age", "3").toList();