+++ lsql-guice/src/test/java/com/w11k/lsql/tests/dao/TestDao.java
@@ -1,12 +1,12 @@
 package com.w11k.lsql.tests.dao;
 
-import com.w11k.lsql.query.RowQuery;
+import com.w11k.lsql.query.PlainQuery;
 import com.w11k.lsql.statement.AbstractSqlStatement;
 import com.w11k.lsql.guice.LSqlDao;
 
 public class TestDao extends LSqlDao {
 
-    public AbstractSqlStatement<RowQuery> getStatement() {
+    public AbstractSqlStatement<PlainQuery> getStatement() {
         return statement("query1");
     }